changed float to double

+++ vertexfinder.c
@@ -2,8 +2,8 @@
 #include <stdlib.h>
 
 
-float findVX(float a, float b) {
-  float vx = -b / (2 * a);
+double findVX(double a, double b) {
+  double vx = -b / (2 * a);
   if (vx == 0){
     return 0;
   }
@@ -12,12 +12,12 @@ float findVX(float a, float b) {
   }
 }
 
-float findY(float a, float b, float c, float x) {
-  float y = (a * x * x) + (b * x) + c;
+double findY(double a, double b, double c, double x) {
+  double y = (a * x * x) + (b * x) + c;
   return y;
 }
 
-void printvform(float a,float vx,float y) {
+void printvform(double a,double vx,double y) {
   //printf("The vertex form is %f(x-%f)+%f\n",a,vx,y);
   printf("The vertex form is ");
   if (a != 1) {
@@ -27,7 +27,7 @@ void printvform(float a,float vx,float y) {
       printf("-%f",vx);
     }
     else if (vx < 0) {
-      float tmp = vx * -1;      
+      double tmp = vx * -1;      
       printf("+%f",tmp);
     }
     printf(")^2");
@@ -42,7 +42,7 @@ void printvform(float a,float vx,float y) {
     }
   }
   else {
-    float tmp = vx + y;
+    double tmp = vx + y;
     if (tmp > 0) {
       printf("x^2+%f\n",tmp);
     }
@@ -56,10 +56,10 @@ void printvform(float a,float vx,float y) {
   }
 }
 
-void rpoints(float vx, float a, float b, float c) {
+void rpoints(double vx, double a, double b, double c) {
   printf("Some points near the vertex are: \n");
-  float tempx;
-  float tempy;
+  double tempx;
+  double tempy;
   for (int i = -3; i <= 3; i++) {
     tempx = vx + i;
     tempy = findY(a,b,c,tempx);
@@ -67,9 +67,9 @@ void rpoints(float vx, float a, float b, float c) {
   }
 }
 
-void doVertex (float a, float b, float c) {
-  float vx = findVX(a,b);
-  float vy = findY(a,b,c,vx);
+void doVertex (double a, double b, double c) {
+  double vx = findVX(a,b);
+  double vy = findY(a,b,c,vx);
   printf("The vertex is (%f, %f)\n", vx, vy);
   printvform(a,vx,vy);
   rpoints(vx,a,b,c);
@@ -77,9 +77,9 @@ void doVertex (float a, float b, float c) {
 
 int main (int argc, char* argv[]) {
   if (argc == 4) {
-    float a = atoi(argv[1]);
-    float b = atoi(argv[2]);
-    float c = atoi(argv[3]);
+    double a = atoi(argv[1]);
+    double b = atoi(argv[2]);
+    double c = atoi(argv[3]);
     
     if (a == 0) {
       printf("Please enter a valid number for each argument.\n");
